refactor(core): remove route group from account pages
