+++ core/app/[locale]/(default)/account/_components/account-notification.tsx
