+++ core/app/[locale]/(default)/account/_components/account-status-provider.tsx
@@ -4,8 +4,6 @@ import { createContext, ReactNode, useContext, useEffect, useState } from 'react
 
 import { usePathname } from '~/i18n/routing';
 
-// import { State as AccountState } from '../settings/change-password/_actions/change-password';
-
 export interface AccountState {
   status: 'idle' | 'error' | 'success';
   messages: string[];
