+++ core/app/[locale]/(default)/account/_components/modal.tsx
