+++ core/app/[locale]/(default)/account/_components/submit-errors-list.tsx
