+++ core/app/[locale]/(default)/account/_components/tab-heading.tsx
