+++ core/app/[locale]/(default)/account/_components/tab-navigation.tsx
