+++ core/app/[locale]/(default)/account/addresses/_actions/delete-address.ts
