+++ core/app/[locale]/(default)/account/addresses/_components/address-book.tsx
