+++ core/app/[locale]/(default)/account/addresses/edit/[slug]/_actions/update-address.ts
