+++ core/app/[locale]/(default)/account/addresses/edit/[slug]/page.tsx
