+++ core/app/[locale]/(default)/account/addresses/page-data.ts
