+++ core/app/[locale]/(default)/account/addresses/page.tsx
