+++ core/app/[locale]/(default)/account/layout.tsx
@@ -1,16 +1,29 @@
-import { getLocale } from 'next-intl/server';
+import { setRequestLocale } from 'next-intl/server';
 import { PropsWithChildren } from 'react';
 
 import { auth } from '~/auth';
 import { redirect } from '~/i18n/routing';
 
-export default async function AccountLayout({ children }: PropsWithChildren) {
-  const locale = await getLocale();
+import { TabNavigation } from './_components/tab-navigation';
+
+interface Props extends PropsWithChildren {
+  params: Promise<{ locale: string }>;
+}
+
+export default async function AccountLayout({ children, params }: Props) {
+  const { locale } = await params;
   const session = await auth();
+
+  setRequestLocale(locale);
 
   if (!session) {
     redirect({ href: '/login', locale });
   }
 
-  return children;
+  return (
+    <>
+      <TabNavigation />
+      {children}
+    </>
+  );
 }
