+++ core/app/[locale]/(default)/account/order/[slug]/_components/order-details.tsx
