+++ core/app/[locale]/(default)/account/order/[slug]/page.tsx
