+++ core/app/[locale]/(default)/account/orders/_components/product-snippet.tsx
