+++ core/app/[locale]/(default)/account/orders/page.tsx
