+++ core/app/[locale]/(default)/account/settings/change-password/_components/change-password-form.tsx
