+++ core/app/[locale]/(default)/account/settings/change-password/page.tsx
