+++ core/app/[locale]/(default)/account/settings/page-data.tsx
