+++ core/app/[locale]/(default)/account/settings/page.tsx
